Suggested Ability to have PlaceHolder
If you want your "input" area to have placeholder text, like a normal input text field allows, this allows for that. When the area is clicked or if there are bit tags, the placeholder is hidden, like a normal placeholder.

NOTE: this is different from the placeholder for the autocomplete, which is a useful, but different feature.