Fix Keys for Mootools 1.4.5
They left, right, delete, and backspace keys were not working when used with Mootools 1.4.5. This upgrades the code to work with the changes to mootools. There may be a better way to do it, but it appeared that the old method for calling the code value for a specific key just didn't work anymore, e.g. `Event.Keys.left` didn't produce the right event.code, so I switched it to event.key and hard coded the values.