Update to Make Up/Down Keys work with Mootools 1.4.5
The old method used `Event.keys.up` to determine the correct `ev.code` for the key. But that doesn't work anymore evidently. So this fix hard codes the key name, and compares it with `ev.key` instead.

NOTE: this may not be the most compatible method, but it was the quickest fix for my purposes and hopefully it will help someone else.